SMARTDocsProject/requirements.txt
django-cleanup
django-extensions == 2.0.7
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pymorphy2-dicts-ru==2.4.404381.4453942
docxtpl==0.4.13
Jinja2==2.10
psycopg2==2.7.5
Pillow==5.2.0